create utils method
